Examine contiguity actions which have been performed on islands by `st_bridges()`

Examine contiguity actions which have been performed on islands by `st_bridges()`

Usage

st_check_islands(data)

Arguments

data

an `sf` dataframe with a neighbourhood column called `"nb"` such as the output of `st_bridges()`.

Value

A dataframe reporting non-contiguous connections made by `st_bridges()`.

Examples

st_bridges(uk_election,"constituency_name") |>
st_check_islands()
#>           island_names island_num nb_num                              nb_names
#> 1        Isle Of Wight        292    379                       New Forest West
#> 2 Na h-Eileanan An Iar        370    461               Ross, Skye and Lochaber
#> 3  Orkney and Shetland        423    101 Caithness, Sutherland and Easter Ross
#> 4             Ynys Mon        630     12                                 Arfon